requirements.txt
streamlit==1.50.0
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.30
langchain-ollama==0.3.8
langchain-google-genai==2.0.10
ddgs==9.6.1
newspaper3k==0.2.8
tqdm==4.67.1
pandas==2.2.3
nltk==3.9.1
faiss-cpu==1.12.0